project func: Add data consistency checking for makefile target: cpi.
After executing the restore checkpoint part of gem5, we enter the automated data comparison phase, and the execution of cpi target will set a judgment condition flag, and set the cpi (0 or stats from gem5) according to the value of the flag.
The specific process is as follows: first determine whether the number of checkpoints is correct, by checking whether the number of CKP files in the m5out directory is the same as the number of lines in the .merge file; then perform the weights check of each checkpoint, by checking whether the weights tag of the ckp file in the m5out directory appears in the .weights file.
If both of the above judgments are true, which means that there is no exception, then the value of CPI will be set normally according to the stats file of gem5, otherwise it will be set to 0.
This process is performed every time gem5 is restore_* or gen_restore_compare_excel.
The goal is to ensure that the M1 and gem5 checkpoints are consistent when automatically comparing data. It is possible that the checkpoint used last time is not the corresponding checkpoint in M1, and the Excel table cannot be compared normally.